requirements.txt
aerich==0.7.2
aiofiles==24.1.0
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.5.0
asyncpg==0.29.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorlog==6.8.2
cryptography==43.0.1
dictdiffer==0.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
iso8601==1.1.0
Jinja2==3.1.4
line_profiler==4.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
rich==13.8.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.38.5
tomlkit==0.13.2
tortoise-orm==0.21.6
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
